+++ itm/tests/singles.rs
@@ -3,7 +3,7 @@ use itm::*;
 #[test]
 fn eof() {
     let empty: &[u8] = &[];
-    let mut decoder = Decoder::new(empty, DecoderOptions { ignore_eof: false });
+    let decoder = Decoder::new(empty, DecoderOptions { ignore_eof: false });
 
     assert!(decoder.singles().next().is_none());
 }
@@ -13,7 +13,7 @@ fn decode_sync_packet() {
     let mut trace_data: Vec<u8> = [0; 47 / 8].to_vec();
     trace_data.push(1 << 7);
 
-    let mut decoder = Decoder::new(trace_data.as_slice(), DecoderOptions { ignore_eof: false });
+    let decoder = Decoder::new(trace_data.as_slice(), DecoderOptions { ignore_eof: false });
     assert_eq!(
         decoder.singles().next().unwrap().unwrap(),
         TracePacket::Sync
@@ -23,7 +23,7 @@ fn decode_sync_packet() {
 #[test]
 fn decode_overflow_packet() {
     let overflow: &[u8] = &[0b0111_0000];
-    let mut decoder = Decoder::new(overflow, DecoderOptions { ignore_eof: false });
+    let decoder = Decoder::new(overflow, DecoderOptions { ignore_eof: false });
     assert_eq!(
         decoder.singles().next().unwrap().unwrap(),
         TracePacket::Overflow
@@ -40,7 +40,7 @@ fn decode_local_timestamp_packets() {
         // LTS2
         0b0101_0000,
     ];
-    let mut decoder = Decoder::new(lts, DecoderOptions { ignore_eof: false });
+    let mut decoder = Decoder::new(lts, DecoderOptions { ignore_eof: false }).singles();
 
     for packet in [
         TracePacket::LocalTimestamp1 {
@@ -52,7 +52,7 @@ fn decode_local_timestamp_packets() {
     .iter()
     .cloned()
     {
-        assert_eq!(decoder.singles().next().unwrap().unwrap(), packet);
+        assert_eq!(decoder.next().unwrap().unwrap(), packet);
     }
 }
 
@@ -80,7 +80,7 @@ fn decode_global_timestamp_packets() {
         0b1111_0100,
         0b0000_0111,
     ];
-    let mut decoder = Decoder::new(gts, DecoderOptions { ignore_eof: false });
+    let mut decoder = Decoder::new(gts, DecoderOptions { ignore_eof: false }).singles();
 
     for packet in [
         TracePacket::GlobalTimestamp1 {
@@ -98,14 +98,14 @@ fn decode_global_timestamp_packets() {
     .iter()
     .cloned()
     {
-        assert_eq!(decoder.singles().next().unwrap().unwrap(), packet);
+        assert_eq!(decoder.next().unwrap().unwrap(), packet);
     }
 }
 
 #[test]
 fn decode_extention_packet() {
     let ext: &[u8] = &[0b0111_1000];
-    let mut decoder = Decoder::new(ext, DecoderOptions { ignore_eof: false });
+    let decoder = Decoder::new(ext, DecoderOptions { ignore_eof: false });
     assert_eq!(
         decoder.singles().next().unwrap().unwrap(),
         TracePacket::Extension { page: 0b111 }
@@ -121,7 +121,7 @@ fn decode_instrumentation_packet() {
         0b0011_1111,
         0b1111_1111,
     ];
-    let mut decoder = Decoder::new(instr, DecoderOptions { ignore_eof: false });
+    let decoder = Decoder::new(instr, DecoderOptions { ignore_eof: false });
 
     assert_eq!(
         decoder.singles().next().unwrap().unwrap(),
@@ -145,7 +145,7 @@ fn decode_eventcounterwrap_packet() {
             0b0000_0101,
             0b0010_1010
         ];
-    let mut decoder = Decoder::new(event, DecoderOptions { ignore_eof: false });
+    let decoder = Decoder::new(event, DecoderOptions { ignore_eof: false });
 
     assert_eq!(
         decoder.singles().next().unwrap().unwrap(),
@@ -168,7 +168,7 @@ fn decode_exceptiontrace_packet() {
             0b0010_0000,
             0b0011_0000
         ];
-    let mut decoder = Decoder::new(excpt, DecoderOptions { ignore_eof: false });
+    let decoder = Decoder::new(excpt, DecoderOptions { ignore_eof: false });
 
     assert_eq!(
         decoder.singles().next().unwrap().unwrap(),
@@ -192,7 +192,7 @@ fn decode_pcsample_packet() {
         0b0001_0101,
         0b0000_0000,
     ];
-    let mut decoder = Decoder::new(samples, DecoderOptions { ignore_eof: false });
+    let mut decoder = Decoder::new(samples, DecoderOptions { ignore_eof: false }).singles();
 
     for packet in [
         TracePacket::PCSample {
@@ -203,7 +203,7 @@ fn decode_pcsample_packet() {
     .iter()
     .cloned()
     {
-        assert_eq!(decoder.singles().next().unwrap().unwrap(), packet);
+        assert_eq!(decoder.next().unwrap().unwrap(), packet);
     }
 }
 
@@ -216,7 +216,7 @@ fn decode_datatracepc_packet() {
         0b0011_1111,
         0b1111_1111,
     ];
-    let mut decoder = Decoder::new(pc, DecoderOptions { ignore_eof: false });
+    let decoder = Decoder::new(pc, DecoderOptions { ignore_eof: false });
 
     assert_eq!(
         decoder.singles().next().unwrap().unwrap(),
@@ -235,7 +235,7 @@ fn decode_datatraceaddress_packet() {
             0b0000_0011,
             0b0000_1111,
         ];
-    let mut decoder = Decoder::new(address, DecoderOptions { ignore_eof: false });
+    let decoder = Decoder::new(address, DecoderOptions { ignore_eof: false });
 
     assert_eq!(
         decoder.singles().next().unwrap().unwrap(),
@@ -267,7 +267,7 @@ fn decode_datatracevalue_packet() {
         0b1010_1101,
         0b0000_0011,
     ];
-    let mut decoder = Decoder::new(payloads, DecoderOptions { ignore_eof: false });
+    let mut decoder = Decoder::new(payloads, DecoderOptions { ignore_eof: false }).singles();
 
     for packet in [
         TracePacket::DataTraceValue {
@@ -302,6 +302,6 @@ fn decode_datatracevalue_packet() {
     .iter()
     .cloned()
     {
-        assert_eq!(decoder.singles().next().unwrap().unwrap(), packet);
+        assert_eq!(decoder.next().unwrap().unwrap(), packet);
     }
 }